[+] add : merge job status to one component + add default job
