+++ pages/carrier/driver/overview.tsx
@@ -9,7 +9,6 @@ import { DRIVER_STATUS_LIST } from '../../../data/carrier'
 import ResourceOverview from '../../../components/common/ResourceOverview'
 import { getDriver } from '../../../components/utilities/apis'
 import { DriverDocument } from '../../../entities/interface/driver'
-import { resourceCreatedState } from '../../../store/atoms/overviewPageState'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import { tableDataState } from '../../../store/atoms/tableState'
 import { alertPropertyState } from '../../../store/atoms/alertPropertyState'
