+++ pages/jobs/add/[step].tsx
@@ -5,49 +5,12 @@ import JobAddStepTwo from '../../../components/switcher/jobs/add/JobAddStepTwo'
 import JobAddStepThree from '../../../components/switcher/jobs/add/JobAddStepThree'
 import JobAddStepFour from '../../../components/switcher/jobs/add/JobAddStepFour'
 import { JobDetails } from '../../../entities/interface/job'
+import { DEFAULT_JOB } from '../../../data/jobs'
 
 const AddJobSwitcherPage = () => {
 	const router = useRouter()
 	const { step } = router.query
-	const [jobDetails, setJobDetails] = useState<JobDetails>({
-		pickup_location: {
-			address: "",
-			district: "",
-			province: "",
-			zipcode: "",
-			latitude: null,
-			longitude: null
-		},
-		dropoff_location: {
-			address: "",
-			district: "",
-			province: "",
-			zipcode: "",
-			latitude: null,
-			longitude: null
-		},
-		pickup_date: new Date,
-		dropoff_date: new Date,
-		weight: undefined,
-		carrier_specification: { 
-			truck: {
-				age: undefined,
-				property: {
-					type: "รถ 4 ล้อ",
-					option: "ตู้ทึบ",
-					chassis: null
-				}
-			},
-			driver: {
-				driver_license_type: "ประเภท 1"
-			}
-		},
-		product_type: "",
-		offer_price: undefined,
-		auto_price: 4000,
-		distance: undefined,
-		permission: "public",
-	})
+	const [jobDetails, setJobDetails] = useState<JobDetails>(DEFAULT_JOB)
 
 	return (
 		<div>
