+++ pages/jobs/index.tsx
@@ -7,8 +7,7 @@ import { PrimaryButton } from "../../components/styles/GlobalComponents"
 import NavigationBar from "../../components/common/NavigationBar"
 import { getAllJobs } from "../../components/utilities/apis"
 import { JobDocument } from "../../entities/interface/job"
-import { useRecoilValue, useSetRecoilState } from "recoil"
-import { resourceCreatedState } from "../../store/atoms/overviewPageState"
+import { useRecoilValue } from "recoil"
 import { alertPropertyState } from "../../store/atoms/alertPropertyState"
 import Alert from "../../components/common/Alert"
 import { userInfoState } from '../../store/atoms/userInfoState'
