--- store/atoms/overviewPageState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const resourceCreatedState = atom({
-	key: 'resourceCreatedStatus',
-	default: "", 
-});
